backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
requests==2.31.0
sqlalchemy==2.0.25
alembic==1.13.1
